UI: Refactor (extract util methods)
