+++ HoloRepositoryUI/src/__tests__/samples/samplePractitioner.json
@@ -10,7 +10,6 @@
   "email": "noreen.spence@nhs.uk",
   "phone": "+1 (876) 557-3574",
   "dateOfBirth": "1969-07-07T00:00:00.000Z",
-  "pictureUrl": "http://placehold.it/64x64",
   "address": {
     "street": "118 Pulaski Street",
     "city": "Rowe",
